refactor / Task <-> Timetable fix error
