+++ app/src/main/java/com/github/monolabhituszi/portal/ui/task/TaskViewModel.kt
@@ -1,13 +1,13 @@
-package com.github.monolabhituszi.portal.ui.task
+package com.github.monolabhituszi.portal.ui.timetable
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class TaskViewModel : ViewModel() {
+class TimetableViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is task Fragment"
+        value = "This is timetable Fragment"
     }
     val text: LiveData<String> = _text
 }
