+++ app/src/main/java/com/github/monolabhituszi/portal/ui/timetable/TimetableAdapter.kt
