+++ app/src/main/java/com/github/monolabhituszi/portal/ui/timetable/TimetableViewModel.kt
@@ -1,13 +1,13 @@
-package com.github.monolabhituszi.portal.ui.timetable
+package com.github.monolabhituszi.portal.ui.task
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class TimetableViewModel : ViewModel() {
+class TaskViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is timetable Fragment"
+        value = "This is task Fragment"
     }
     val text: LiveData<String> = _text
 }
