+++ app/src/main/res/layout/fragment_timetable.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.timetable.TimetableFragment">
+        tools:context=".ui.task.TaskFragment">
 
         <TextView
             android:id="@+id/text_dashboard"
